Flav/views only for managed datasources (#6608)
* s/datasource/data_source

* Only create a data source view for managed data sources with a connection

* Create data source with provider

* Add migration to delete un managed data source views

* :see_no_evil:

* :sparkles:

* :sparkles:

+++ front/lib/resources/data_source_resource.ts
@@ -163,7 +163,9 @@ export class DataSourceResource extends BaseResource<DataSource> {
     auth: Authenticator,
     transaction?: Transaction
   ): Promise<Result<undefined, Error>> {
-    await DataSourceViewResource.deleteForDataSource(auth, this);
+    if (this.isManaged()) {
+      await DataSourceViewResource.deleteForDataSource(auth, this);
+    }
 
     try {
       await this.model.destroy({
@@ -211,6 +213,16 @@ export class DataSourceResource extends BaseResource<DataSource> {
       },
     };
   }
+
+  isManaged(): boolean {
+    return (
+      this.name.startsWith("managed-") &&
+      this.connectorProvider !== null &&
+      this.connectorProvider !== "webcrawler"
+    );
+  }
+
+  // Serialization.
 
   toJSON(): DataSourceType {
     return {
